remove a -Wconversion warning

+++ lib/sfio/sfnew.c
@@ -89,7 +89,7 @@ Sfio_t *sfnew(Sfio_t * oldf, void * buf, size_t size, int file,
 
     /* stream type */
     f->mode = (flags & SF_READ) ? SF_READ : SF_WRITE;
-    f->flags = (flags & SF_FLAGS) | (sflags & (SF_MALLOC | SF_STATIC));
+    f->flags = (unsigned short)((flags & SF_FLAGS) | (sflags & (SF_MALLOC | SF_STATIC)));
     f->bits = (flags & SF_RDWR) == SF_RDWR ? SF_BOTH : 0;
     f->file = file;
     f->here = f->extent = 0;
